refactor: share one Robolectric harness across the reader tests
Three test classes carried their own copy of the same four helpers: reset
the singleton, listen to reader status, mount a widget on a Robolectric
activity, and mock a publication that conforms to one profile. Phase 2 of
the coroutine-handler work adds more classes that need them, so the
copies move into ReaderTestHarness.kt next to ReadiumReaderFields.kt,
which already owns the reflection into the singleton.

The harness also gains an error-channel listener, and the mount helper is
named buildReaderWidget now: what it builds depends on the publication
seeded before the call, not on the helper.

No test case changed.
