+++ flureadium/android/src/test/kotlin/dev/mulev/flureadium/ReadiumReaderWidgetAudioHostTest.kt
@@ -1,22 +1,15 @@
 package dev.mulev.flureadium
 
-import android.content.ContextWrapper
 import android.os.Build
-import androidx.fragment.app.FragmentActivity
-import dev.mulev.flureadium.events.ReaderStatusEventChannel
-import io.flutter.plugin.common.BinaryMessenger
-import io.flutter.plugin.common.EventChannel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
 import org.junit.runner.RunWith
-import org.mockito.Mockito.`when`
 import org.mockito.Mockito.mock
 import org.readium.r2.shared.ExperimentalReadiumApi
 import org.readium.r2.shared.publication.Publication
-import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 import kotlin.test.AfterTest
@@ -31,11 +24,10 @@ import kotlin.test.assertSame
  * mounts with no visual navigator, and its teardown releases only what it
  * registered.
  *
- * Uses reflection to set private fields on the ReadiumReader singleton, the
- * same harness as ReadiumReaderSamePubCacheTest. Dispatchers.setMain with an
- * unconfined test dispatcher makes the widget's init coroutine run inline
- * during construction, so both assertions hold as soon as the constructor
- * returns.
+ * Seeds the ReadiumReader singleton and mounts the widget through
+ * ReaderTestHarness.kt. Dispatchers.setMain with an unconfined test dispatcher
+ * makes the widget's init coroutine run inline during construction, so both
+ * assertions hold as soon as the constructor returns.
  */
 @RunWith(RobolectricTestRunner::class)
 @Config(sdk = [Build.VERSION_CODES.P], manifest = Config.NONE)
@@ -48,7 +40,7 @@ internal class ReadiumReaderWidgetAudioHostTest {
     fun setUp() {
         Dispatchers.setMain(testDispatcher)
         resetReaderState()
-        setReaderField("_currentPublication", audiobookPublication())
+        setReaderField("_currentPublication", publicationConformingTo(Publication.Profile.AUDIOBOOK))
     }
 
     @AfterTest
@@ -57,23 +49,16 @@ internal class ReadiumReaderWidgetAudioHostTest {
         resetReaderState()
     }
 
-    private fun resetReaderState() {
-        setReaderField("_currentPublication", null)
-        setReaderField("isReadyEventChannel", null)
-        setReaderField("readerStatusEventChannel", null)
-        ReadiumReader.currentReaderWidget = null
-    }
-
     @Test
     fun audioHostDoesNotSetUpEpubNavigatorMachinery() {
-        buildAudioHostWidget()
+        buildReaderWidget()
 
         assertNull(getReaderField("isReadyEventChannel"))
     }
 
     @Test
     fun audioHostDisposeKeepsNewerWidgetRegistration() {
-        val audioWidget = buildAudioHostWidget()
+        val audioWidget = buildReaderWidget()
         val newerWidget = mock(ReadiumReaderWidget::class.java)
         ReadiumReader.currentReaderWidget = newerWidget
 
@@ -86,53 +71,18 @@ internal class ReadiumReaderWidgetAudioHostTest {
     fun audioHostReportsLoadingThenReady() {
         val statuses = subscribeToReaderStatus()
 
-        buildAudioHostWidget()
+        buildReaderWidget()
 
         assertEquals(listOf("loading", "ready"), statuses)
     }
 
     @Test
     fun activeAudioHostDisposeClearsItsRegistration() {
-        val audioWidget = buildAudioHostWidget()
+        val audioWidget = buildReaderWidget()
 
         audioWidget.dispose()
 
         assertNull(ReadiumReader.currentReaderWidget)
     }
 
-    /** Attaches a listener to the reader-status channel and records what it receives. */
-    private fun subscribeToReaderStatus(): List<String> {
-        val received = mutableListOf<String>()
-        val channel = ReaderStatusEventChannel(mock(BinaryMessenger::class.java))
-        setReaderField("readerStatusEventChannel", channel)
-        channel.onListen(
-            null,
-            object : EventChannel.EventSink {
-                override fun success(event: Any?) {
-                    received.add(event as String)
-                }
-
-                override fun error(errorCode: String, errorMessage: String?, errorDetails: Any?) = Unit
-
-                override fun endOfStream() = Unit
-            }
-        )
-        return received
-    }
-
-    private fun buildAudioHostWidget(): ReadiumReaderWidget {
-        val activity = Robolectric.buildActivity(FragmentActivity::class.java).setup().get()
-        return ReadiumReaderWidget(
-            ContextWrapper(activity),
-            1,
-            emptyMap(),
-            mock(BinaryMessenger::class.java)
-        )
-    }
-
-    private fun audiobookPublication(): Publication {
-        val publication = mock(Publication::class.java)
-        `when`(publication.conformsTo(Publication.Profile.AUDIOBOOK)).thenReturn(true)
-        return publication
-    }
 }
